requirements.txt
requests
discord.py
python-dotenv
google-generativeai
motor